RavenDB-17340 - Addressed PR comments and failing test
